agent.py
# agent.py

import asyncio
from dataclasses import dataclass, field
import uuid
from typing import Any, Dict, Optional
from langchain_core.prompts import ChatPromptTemplate
from langchain_core.messages import SystemMessage, HumanMessage, AIMessage

# Import enhancement systems
from python.systems.pattern_recognition import PatternEngine
from python.systems.resource_optimization import ResourceManager
from python.systems.learning_integration import LearningEngine
from python.systems.multi_agent_cooperation import CoordinationEngine
from python.systems.system_evolution import EvolutionEngine
from python.systems.analytics_system import AnalyticsEngine
from python.systems.interface_enhancement import InterfaceEngine

@dataclass
class ModelConfig:
    provider: str
    name: str
    ctx_length: int
    limit_requests: int
    limit_input: int
    limit_output: int
    kwargs: Dict[str, Any] = field(default_factory=dict)

@dataclass
class AgentConfig:
    # Core configurations
    chat_model: ModelConfig
    utility_model: ModelConfig
    embeddings_model: ModelConfig
    prompts_subdir: str = ""
    memory_subdir: str = ""
    
    # Enhancement systems configurations
    pattern_recognition: dict = field(default_factory=lambda: {
        'similarity_threshold': 0.75,
        'min_examples': 5,
        'learning_rate': 0.1
    })
    
    resource_optimization: dict = field(default_factory=lambda: {
        'optimization_threshold': 0.8,
        'allocation_strategy': 'dynamic'
    })
    
    learning_integration: dict = field(default_factory=lambda: {
        'experience_threshold': 0.7,
        'learning_rate': 0.1
    })
    
    multi_agent_cooperation: dict = field(default_factory=lambda: {
        'coordination_threshold': 0.8,
        'team_size': 3
    })
    
    system_evolution: dict = field(default_factory=lambda: {
        'evolution_rate': 0.2,
        'adaptation_threshold': 0.7
    })
    
    analytics_system: dict = field(default_factory=lambda: {
        'analysis_window': 1000,
        'confidence_threshold': 0.8
    })
    
    interface_system: dict = field(default_factory=lambda: {
        'response_format': 'markdown',
        'style_preferences': {}
    })

@dataclass
class AgentContext:
    _contexts: Dict[str, "AgentContext"] = field(default_factory=dict)
    _counter: int = 0
    id: str = field(default_factory=lambda: str(uuid.uuid4()))
    name: Optional[str] = None
    config: AgentConfig = field(default_factory=lambda: AgentConfig(**create_default_config()))
    log: Any = field(default_factory=lambda: Log())
    agent0: Optional["Agent"] = None
    paused: bool = False
    streaming_agent: Optional["Agent"] = None
    process: Any = None
    no: int = field(init=False)

    def __post_init__(self):
        if self.agent0 is None:
            self.agent0 = Agent(0, self.config, self)
        AgentContext._counter += 1
        self.no = AgentContext._counter
        
        # Store context
        existing = self._contexts.get(self.id)
        if existing:
            AgentContext.remove(self.id)
        self._contexts[self.id] = self

    @classmethod
    def get(cls, id: str) -> Optional["AgentContext"]:
        return cls._contexts.get(id)

    @classmethod
    def remove(cls, id: str):
        if id in cls._contexts:
            del cls._contexts[id]

class Agent:
    def __init__(self, number: int, config: AgentConfig, context: AgentContext | None = None):
        # Core initialization
        self.number = number
        self.config = config
        self.context = context or AgentContext(config)
        self.agent_name = f"Agent {self.number}"
        
        # Enhancement systems
        self.pattern_engine = PatternEngine(self.config.pattern_recognition)
        self.resource_manager = ResourceManager(self.config.resource_optimization)
        self.learning_engine = LearningEngine(self.config.learning_integration)
        self.cooperation_engine = CoordinationEngine(self.config.multi_agent_cooperation)
        self.evolution_engine = EvolutionEngine(self.config.system_evolution)
        self.analytics_engine = AnalyticsEngine(self.config.analytics_system)
        self.interface_engine = InterfaceEngine(self.config.interface_system)
        
        # State management
        self.history = []
        self.last_user_message = None
        self.intervention = None
        self.data = {}

    async def monologue(self):
        while True:
            try:
                # Initialize loop
                self.loop_data = self.initialize_loop()
                
                # Process through enhancement systems
                await self.process_through_systems()
                
                # Main conversation loop
                while True:
                    try:
                        # Prepare prompt
                        prompt = await self.prepare_prompt()
                        
                        # Generate response
                        response = await self.generate_response(prompt)
                        
                        # Process response
                        result = await self.process_response(response)
                        
                        if result:
                            return result
                            
                    except Exception as e:
                        await self.handle_exception(e)
                        
            except Exception as e:
                self.handle_critical_exception(e)

    async def process_through_systems(self):
        # Pattern recognition
        patterns = await self.pattern_engine.process_patterns(self.loop_data)
        
        # Resource optimization
        resources = await self.resource_manager.optimize_resources()
        
        # Learning integration
        learning = await self.learning_engine.process_learning()
        
        # Multi-agent cooperation
        cooperation = await self.cooperation_engine.coordinate_agents()
        
        # System evolution
        evolution = await self.evolution_engine.evolve_system()
        
        # Analytics processing
        analytics = await self.analytics_engine.analyze_system()
        
        # Interface enhancement
        interface = await self.interface_engine.process_interaction()
        
        # Integrate results
        self.integrate_system_results(
            patterns, resources, learning, cooperation,
            evolution, analytics, interface
        )

    async def prepare_prompt(self):
        # Get system prompt
        system_prompt = await self.get_system_prompt()
        
        # Get enhancement data
        system_data = await self.get_system_enhancements()
        
        # Combine prompts
        prompt = ChatPromptTemplate.from_messages([
            SystemMessage(content="\n\n".join(system_prompt + system_data)),
            *self.format_history()
        ])
        
        return prompt

    async def generate_response(self, prompt):
        # Process through interface
        formatted_prompt = await self.interface_engine.format_prompt(prompt)
        
        # Generate response
        response = await self.call_model(formatted_prompt)
        
        # Analyze response
        await self.analytics_engine.analyze_response(response)
        
        return response

    async def process_response(self, response):
        # Process through interface
        formatted_response = await self.interface_engine.format_response(response)
        
        # Extract and execute tools
        result = await self.process_tools(formatted_response)
        
        # Learn from execution
        await self.learning_engine.learn_from_execution(result)
        
        return result

    # Additional helper methods...

class Log:
    def __init__(self):
        self.entries = []

    def add(self, entry):
        self.entries.append(entry)

def create_default_config():
    return {
        'chat_model': {
            'provider': 'openai',
            'name': 'gpt-4',
            'ctx_length': 2048,
            'limit_requests': 100,
            'limit_input': 2048,
            'limit_output': 2048,
            'kwargs': {
                'temperature': 0.7
            }
        },
        'utility_model': {
            'provider': 'openai',
            'name': 'gpt-3.5-turbo',
            'ctx_length': 2048,
            'limit_requests': 100,
            'limit_input': 2048,
            'limit_output': 2048,
            'kwargs': {
                'temperature': 0.3
            }
        },
        'embeddings_model': {
            'provider': 'openai',
            'name': 'text-embedding-ada-002',
            'ctx_length': 2048,
            'limit_requests': 100,
            'limit_input': 2048,
            'limit_output': 2048,
            'kwargs': {}
        },
        'pattern_recognition': {
            'similarity_threshold': 0.75,
            'min_examples': 5,
            'learning_rate': 0.1
        },
        'resource_optimization': {
            'optimization_threshold': 0.8,
            'allocation_strategy': 'dynamic'
        },
        'learning_integration': {
            'experience_threshold': 0.7,
            'learning_rate': 0.1
        },
        'multi_agent_cooperation': {
            'coordination_threshold': 0.8,
            'team_size': 3
        },
        'system_evolution': {
            'evolution_rate': 0.2,
            'adaptation_threshold': 0.7
        },
        'analytics_system': {
            'analysis_window': 1000,
            'confidence_threshold': 0.8
        },
        'interface_system': {
            'response_format': 'markdown',
            'style_preferences': {}
        }
    }